Try to encode passed form values as UTF-8 before sending it to Akismet
If any non-ASCII text is submitted in the form, urlencode() would encode the string implicitly to ASCII
which fails if there are non-ASCII characters in the string. And since just "Exception" is caught and
ignored silently, those submits would never passed to Akismet and so not detected as spam.
If the encoding to UTF-8 fails, we still try to pass it as is to Akismet, like before.