Use CSS instead of JavaScript for the admin app dropdowns
There were some bugs with the dropdowns when they were in JavaScript:

  1.  When you open a dropdown and then scroll, the dropdown would stay
      put, instead of following the scroll.

  2.  The JavaScript used `.live('mouseover')` which binds to body and
      wastes memory because it's fired for mouseover on every single DOM
      element.

  3.  Occasionally, the dropdowns never disappeared even after mouseout.

This commit fixes those bugs by using CSS and `:hover` instead of
JavaScript.  Additionally, it simplifies the JavaScript related to
setting the href of the primary menu item links to their first child.

It is a pixel for pixel match of the previous functionality and
appearance.